Separate protobuf encoding from the parquet server
