+++ persistence-protobuf/src/main/java/io/mishmash/opentelemetry/persistence/proto/package-info.java
@@ -1,0 +1,37 @@
+/*
+ *    Copyright 2024 Mishmash IO UK Ltd.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+/**
+ * <h2>Flattened protobuf format for OpenTelemetry signals</h2>
+ * <p>
+ * This package contains the classes used by the embedded OpenTelemetry
+ * servers to persist 'flattened' telemetry records. Find out more about the
+ * embedded servers (and how they 'flatten' OpenTelemetry data)
+ * on the GitHub link below.
+ * </p>
+ * <p>
+ * This package is part of a broader set of OpenTelemetry-related activities
+ * at mishmash io. To find out more about this and other packages visit the
+ * links below.
+ * </p>
+ * @see <a href="https://mishmash.io/open_source/opentelemetry">
+ * OpenTelemetry at mishmash io</a>
+ * @see <a href="https://github.com/mishmash-io/opentelemetry-server-embedded">
+ * This package on GitHub</a>
+ * @see <a href="https://mishmash.io/">mishmash io</a>
+ */
+package io.mishmash.opentelemetry.persistence.proto;
