+++ persistence-protobuf/src/main/proto/io/mishmash/opentelemetry/persistence/proto/v1/logs_persistence.proto
@@ -16,15 +16,15 @@
 
 syntax = "proto3";
 
-package io.mishmash.opentelemetry.server.parquet.persistence.proto.v1;
+package io.mishmash.opentelemetry.persistence.proto.v1;
 
 import "opentelemetry/proto/common/v1/common.proto";
-import "opentelemetry/proto/trace/v1/trace.proto";
+import "opentelemetry/proto/logs/v1/logs.proto";
 
-option java_package = "io.mishmash.opentelemetry.server.parquet.persistence.proto.v1";
-option java_outer_classname = "TracesPersistenceProto";
+option java_package = "io.mishmash.opentelemetry.persistence.proto.v1";
+option java_outer_classname = "LogsPersistenceProto";
 
-message PersistedSpan {
+message PersistedLog {
   // OTLP messages are essentially collections of entries (logs, metrics or traces) and a batch is 
   // one such collection. For each of these 'batches' (or messages) a timestamp (batch_timestamp) 
   // is recorded - noting when it was received, and also an UUID is generated and written to batch_UUID.
@@ -51,32 +51,37 @@ message PersistedSpan {
   repeated .opentelemetry.proto.common.v1.KeyValue scope_attributes = 9;
   uint32 scope_dropped_attributes_count = 10;
 
-  // the following are extracted from opentelemetry.proto.trace.v1.Span
-  bytes trace_id = 11;
-  bytes span_id = 12;
-  string trace_state = 13;
-  bytes parent_span_id = 14;
-  fixed32 flags = 15;
-  string name = 16;
-  .opentelemetry.proto.trace.v1.Span.SpanKind kind = 17;
-  fixed64 start_time_unix_nano = 18;
-  fixed64 end_time_unix_nano = 19;
-  repeated .opentelemetry.proto.common.v1.KeyValue attributes = 20;
-  uint32 dropped_attributes_count = 21;
-  repeated .opentelemetry.proto.trace.v1.Span.Event events = 22;
-  uint32 dropped_events_count = 23;
-  repeated .opentelemetry.proto.trace.v1.Span.Link links = 24;
-  uint32 dropped_links_count = 25;
-  .opentelemetry.proto.trace.v1.Status status = 26;
+  // the following are extracted from opentelemetry.proto.logs.v1.LogRecord
+  fixed64 time_unix_nano = 11;
+  fixed64 observed_time_unix_nano = 12;
+  .opentelemetry.proto.logs.v1.SeverityNumber severity_number = 13;
+  string severity_text = 14;
+  
+  // Tells what is the type of the data in the log body - STRING_VALUE, INT_VALUE, BOOL_VALUE, etc
+  string body_type = 15;
+  // depending on body type - use one of these to get the actual log record body:
+  optional string body_string = 16;
+  optional bool body_bool = 17;
+  optional int64 body_int = 18;
+  optional double body_double = 19;
+  optional .opentelemetry.proto.common.v1.ArrayValue body_array = 20;
+  optional .opentelemetry.proto.common.v1.KeyValueList body_kvlist = 21;
+  optional bytes body_bytes = 22;
+
+  repeated .opentelemetry.proto.common.v1.KeyValue attributes = 23;
+  uint32 dropped_attributes_count = 24;
+  fixed32 flags = 25;
+  bytes trace_id = 26;
+  bytes span_id = 27;
 
   // The Schema URL, if known. This is the identifier of the Schema that the log data
   // is recorded in. To learn more about Schema URL see
   // https://opentelemetry.io/docs/specs/otel/schemas/#schema-url
-  string span_schema_url = 27;
+  string log_schema_url = 28;
 
   // is_valid will be false if the system found this record to not comply with OTLP specs.
-  bool is_valid = 28;
+  bool is_valid = 29;
 
   // when is_valid is false, error_message will provide some information on what was wrong.
-  string error_message = 29;
+  string error_message = 30;
 }
