+++ persistence-protobuf/src/main/proto/io/mishmash/opentelemetry/persistence/proto/v1/metrics_persistence.proto
@@ -16,12 +16,12 @@
 
 syntax = "proto3";
 
-package io.mishmash.opentelemetry.server.parquet.persistence.proto.v1;
+package io.mishmash.opentelemetry.persistence.proto.v1;
 
 import "opentelemetry/proto/common/v1/common.proto";
 import "opentelemetry/proto/metrics/v1/metrics.proto";
 
-option java_package = "io.mishmash.opentelemetry.server.parquet.persistence.proto.v1";
+option java_package = "io.mishmash.opentelemetry.persistence.proto.v1";
 option java_outer_classname = "MetricsPersistenceProto";
 
 message PersistedMetric {
