+++ pom.xml
@@ -77,6 +77,7 @@
 
   <modules>
     <module>collector-embedded</module>
+    <module>persistence-protobuf</module>
     <module>server-parquet</module>
   </modules>
